Rename begin/solution to Begin/Solution
