+++ Exercises/fortran/01/Solution/exercise.f90
@@ -1,3 +1,16 @@
+! ************************************************************************
+!
+!                        Kokkos v. 4.0
+!       Copyright (2022) National Technology & Engineering
+!               Solutions of Sandia, LLC (NTESS).
+!
+! Under the terms of Contract DE-NA0003525 with NTESS,
+! the U.S. Government retains certain rights in this software.
+!
+! Part of Kokkos, under the Apache License v2.0 with LLVM Exceptions.
+! See https://kokkos.org/LICENSE for license information.
+! SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+
 program example_axpy_view
   use, intrinsic :: iso_c_binding
   use, intrinsic :: iso_fortran_env
